+++ src/components/search_bar/filters/field_value_toggle_group_filter.js
@@ -69,6 +69,7 @@ export class FieldValueToggleGroupFilter extends Component {
           hasActiveFilters={active}
           noDivider={!isLastItem}
           aria-pressed={active ? 'true' : 'false'}
+          withNext={!isLastItem}
         >
           {name}
         </EuiFilterButton>