+++ src/components/search_bar/search_bar.js
@@ -158,7 +158,7 @@ export class EuiSearchBar extends Component {
     const toolsLeftEl = this.renderTools(toolsLeft);
 
     const filtersBar = !filters ? undefined : (
-      <EuiFlexItem grow={false}>
+      <EuiFlexItem className="euiSearchBar__filtersHolder" grow={false}>
         <EuiSearchFilters filters={filters} query={query} onChange={this.onFiltersChange} />
       </EuiFlexItem>
     );
@@ -166,9 +166,9 @@ export class EuiSearchBar extends Component {
     const toolsRightEl = this.renderTools(toolsRight);
 
     return (
-      <EuiFlexGroup gutterSize="m" alignItems="center">
+      <EuiFlexGroup gutterSize="m" alignItems="center" wrap>
         {toolsLeftEl}
-        <EuiFlexItem grow={true}>
+        <EuiFlexItem className="euiSearchBar__searchHolder" grow={true}>
           <EuiSearchBox
             {...box}
             query={queryText}